server/src/main/java/com/fhict/server/animals/Reservor.java
package com.fhict.server.animals;

import java.time.LocalDateTime;

public class Reservor {

    private String name;
    private LocalDateTime reservedAt;

    public Reservor(String name, LocalDateTime reservedAt) {
        this.name = name;
        this.reservedAt = reservedAt;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public LocalDateTime getReservedAt() {
        return reservedAt;
    }

    public void setReservedAt(LocalDateTime reservedAt) {
        this.reservedAt = reservedAt;
    }
}

