server/src/test/java/com/fhict/server/animals/ReservorTest.java
package com.fhict.server.animals;

import com.fhict.server.animals.Reservor;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.time.LocalDateTime;

import static org.junit.jupiter.api.Assertions.*;
import static org.junit.jupiter.api.Assertions.assertFalse;

public class ReservorTest {

    @Test
    public void TestConstructor()
    {
        LocalDateTime reservedAt = LocalDateTime.now();
        Reservor reservor = new Reservor("John Doe", reservedAt);
        assertEquals("John Doe", reservor.getName());
        assertEquals(reservedAt, reservor.getReservedAt());
    }
}
